+++ login-workflow/src/screens/subScreens/ContactSupportSubscreen/index.ts
@@ -1,2 +1,1 @@
-export * from './ViewEulaSubscreen';
 export * from './ContactSupportSubscreen';