+++ login-workflow/CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v3.2.0 (Unreleased)
+## v4.0.0 (Unreleased)
 
 ### Added
 
+-   A `ViewEulaSubscreen` component.
+-   A `ContactSupportSubscreen` component.
 -   A `BrandedCardContainer` component.
 -   Confirm password field error state message in `ChangePasswordForm`.
+-   Option to hide the Create Password screen during registration
 
 ### Changed